feat: add beam telemetry events
This commit makes it possible to attatch code to the beam telemetry
events that are described in the `wolff_metrics` module. Beam telemetry is
a library for exposing telemetry events to users. This commit also adds
tests for telemetry events and a description of how to use them to the
README file.