refactor: add `gauge_set`/`gauge_shift` APIs
Handling gauge metrics can be hard when restarting producers and lead
to potentially inconsistent metrics.  Here, we refactor the gauge
metrics APIs to be able to set them to their current value, as well
as shifting them as before, plus we add the partition number as a
metadata.  This'll help keep metrics more consistent, and the extra
metadata helps in the case a single metric (e.g.: `batching`) is
shared by multiple producers/workers.

Also, we change the semantics of the `queuing` metrics to represent
the number of _requests_ being queued, each request being a batch of
one or more Kafka messages.