refactor: use shared ets table for partition-producers
Prior to this change, a unique atom is required to start
supervised producers.
This PR changes to use a shared ETS table for partition-producer.
To be backward compabtible, the old 'name' option in producer
config is still required to be used as the namespace for
partitions in the shared table.