fix: unexpected_id crash
Introduced in 4.0.1, the sequential ID assertion only applies to
the backlog items, but not for the inflight items.
Because backlog queue head can be dropped when overloaded,
if the inflight is not empty, when drop happened, the next
insertion to inflight will detect a gap and crash.

This commit allows the gap for inflight queue.